+++ CHANGELOG.md
@@ -33,6 +33,10 @@
 
    Since the script is placed in the end of the HTML, when the DOM is loaded, and JSPM anyway wraps content of each module into a function and leaking into a global scope variables no longer possessing any threat, there is simply no sense to continue doing this in 2018.
 
+- [modules] `node-gettext` has been updated to the version `2.0.0` and related `gettext` module completely rewritten. It is no longer a class, but a simple function, which returns same Gettext methods as before.
+
+   Closes [#156](https://github.com/LotusTM/Kotsu/pull/156), fixes [#272](https://github.com/LotusTM/Kotsu/issues/272), related to [#45](https://github.com/LotusTM/Kotsu/issues/45).
+
 - [modules][grunt][data] Matter data is no longer defined in Nunjucks task and instead passed explicitly as a part of `SITE.matter` data within data `index` file.
 - [data] Matter data property renamed from `SITE._matter` to `SITE.matter`.
 - [data] Data file now destructuring `config()` values once to avoid numerous `config()` calls. This should improve performance in large data files.